requirements.txt
# Core Framework
streamlit>=1.30.0
python-dotenv>=1.0.0

# LLM Service (Free Tier)
google-generativeai>=0.3.0

# RAG Components
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0

# PDF Generation
reportlab>=4.0.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
